Fly!
New Plugin!
Fly is a plugin that allows players to fly around in survival, when a
player takes damage they fall to the ground (without getting kicked of
course). This plugin would be useful as a VIP perk for factions servers
or, could be integrated into a spectator mode for any sort of server.

I hope you guys enjoy this new plugin :)